feat(community): 将社区信号代码从 koduck-core 迁移到 koduck-community 模块
- 迁移 6 个 Entity 类到 koduck-community
- 迁移 6 个 Repository 类到 koduck-community
- 迁移 8 个 DTO 类到 koduck-community
- 迁移 CommunitySignalService 接口及实现到 koduck-community
- 迁移 CommunitySignalController 到 koduck-community
- 迁移 CommunitySignalResponseAssembler 到 koduck-community
- 迁移相关测试文件到 koduck-community
- 创建 CommunityEntityCopyUtils 处理 community 实体的复制逻辑
- 从 koduck-core 移除 community 相关代码

Closes #533
