+++ koduck-backend/koduck-community/src/main/java/com/koduck/entity/community/SignalSubscription.java
@@ -16,6 +16,7 @@ import org.hibernate.annotations.CreationTimestamp;
 
 import com.koduck.entity.auth.User;
 import com.koduck.util.EntityCopyUtils;
+import com.koduck.util.CommunityEntityCopyUtils;
 
 import lombok.AccessLevel;
 import lombok.Data;
@@ -23,193 +24,181 @@ import lombok.NoArgsConstructor;
 import lombok.Setter;
 
 /**
- * 信号点赞实体。
+ * 表示用户对交易信号的订阅实体。
+ * <p>
+ * 存储用户与他们订阅的信号之间的关系，
+ * 以及通知偏好设置。
  *
  * @author Koduck Team
  */
 @Entity
-@Table(name = "signal_likes")
+@Table(name = "signal_subscriptions")
 @Data
 @NoArgsConstructor
-public class SignalLike {
+public class SignalSubscription {
 
-    /** 点赞 ID。 */
+    /**
+     * 订阅的唯一标识符。
+     */
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
     @Setter(AccessLevel.NONE)
     private Long id;
 
-    /** 信号 ID。 */
+    /**
+     * 被订阅信号的标识符。
+     */
     @Column(name = "signal_id", nullable = false)
     private Long signalId;
 
-    /** 用户 ID。 */
+    /**
+     * 订阅用户的标识符。
+     */
     @Column(name = "user_id", nullable = false)
     private Long userId;
 
-    /** 创建时间戳。 */
+    /**
+     * 指示此订阅是否启用通知的标志。
+     */
+    @Column(name = "notify_enabled")
+    private Boolean notifyEnabled = true;
+
+    /**
+     * 订阅创建时间戳。
+     */
     @CreationTimestamp
     @Column(name = "created_at", updatable = false)
     @Setter(AccessLevel.NONE)
     private LocalDateTime createdAt;
 
-    /** 关联的信号。 */
+    /**
+     * 被订阅的信号。
+     */
     @ManyToOne(fetch = FetchType.LAZY)
     @JoinColumn(name = "signal_id", insertable = false, updatable = false)
     private CommunitySignal signal;
 
-    /** 关联的用户。 */
+    /**
+     * 订阅用户。
+     */
     @ManyToOne(fetch = FetchType.LAZY)
     @JoinColumn(name = "user_id", insertable = false, updatable = false)
     private User user;
 
     /**
-     * 获取 SignalLike 的构建器。
+     * 创建 SignalSubscription 的新构建器实例。
      *
-     * @return 构建器
+     * @return 新的构建器实例
      */
     public static Builder builder() {
         return new Builder();
     }
 
     /**
-     * SignalLike 的构建器。
+     * 构建 SignalSubscription 实例的构建器类。
      */
     public static final class Builder {
 
-        /** 点赞 ID。 */
+        /**
+         * 订阅 ID。
+         */
         private Long id;
 
-        /** 信号 ID。 */
+        /**
+         * 信号 ID。
+         */
         private Long signalId;
 
-        /** 用户 ID。 */
+        /**
+         * 用户 ID。
+         */
         private Long userId;
 
-        /** 创建时间戳。 */
-        private LocalDateTime createdAt;
-
-        /** 关联的信号。 */
-        private CommunitySignal signal;
-
-        /** 关联的用户。 */
-        private User user;
+        /**
+         * 通知启用标志。
+         */
+        private Boolean notifyEnabled;
 
         /**
-         * 设置 ID。
-         *
-         * @param id 点赞 ID
-         * @return 构建器
+         * 创建时间戳。
          */
+        private LocalDateTime createdAt;
+
+        /**
+         * 关联的信号。
+         */
+        private CommunitySignal signal;
+
+        /**
+         * 关联的用户。
+         */
+        private User user;
+
         public Builder id(Long id) {
             this.id = id;
             return this;
         }
 
-        /**
-         * 设置信号 ID。
-         *
-         * @param signalId 信号 ID
-         * @return 构建器
-         */
         public Builder signalId(Long signalId) {
             this.signalId = signalId;
             return this;
         }
 
-        /**
-         * 设置用户 ID。
-         *
-         * @param userId 用户 ID
-         * @return 构建器
-         */
         public Builder userId(Long userId) {
             this.userId = userId;
             return this;
         }
 
-        /**
-         * 设置创建时间戳。
-         *
-         * @param createdAt 创建时间戳
-         * @return 构建器
-         */
+        public Builder notifyEnabled(Boolean notifyEnabled) {
+            this.notifyEnabled = notifyEnabled;
+            return this;
+        }
+
         public Builder createdAt(LocalDateTime createdAt) {
             this.createdAt = createdAt;
             return this;
         }
 
-        /**
-         * 设置信号。
-         *
-         * @param signal 信号
-         * @return 构建器
-         */
         public Builder signal(CommunitySignal signal) {
-            this.signal = EntityCopyUtils.copyCommunitySignal(signal);
+            this.signal = CommunityEntityCopyUtils.copyCommunitySignal(signal);
             return this;
         }
 
-        /**
-         * 设置用户。
-         *
-         * @param user 用户
-         * @return 构建器
-         */
         public Builder user(User user) {
             this.user = EntityCopyUtils.copyUser(user);
             return this;
         }
 
         /**
-         * 构建 SignalLike。
+         * 构建新的 SignalSubscription 实例。
          *
-         * @return SignalLike
+         * @return 构建的 SignalSubscription
          */
-        public SignalLike build() {
-            SignalLike like = new SignalLike();
-            like.id = id;
-            like.setSignalId(signalId);
-            like.setUserId(userId);
-            like.createdAt = createdAt;
-            like.setSignal(signal);
-            like.setUser(user);
-            return like;
+        public SignalSubscription build() {
+            SignalSubscription subscription = new SignalSubscription();
+            subscription.id = id;
+            subscription.setSignalId(signalId);
+            subscription.setUserId(userId);
+            subscription.setNotifyEnabled(notifyEnabled);
+            subscription.createdAt = createdAt;
+            subscription.setSignal(signal);
+            subscription.setUser(user);
+            return subscription;
         }
     }
 
-    /**
-     * 获取带有防御性副本的信号。
-     *
-     * @return 信号
-     */
     public CommunitySignal getSignal() {
-        return EntityCopyUtils.copyCommunitySignal(signal);
+        return CommunityEntityCopyUtils.copyCommunitySignal(signal);
     }
 
-    /**
-     * 使用防御性副本设置信号。
-     *
-     * @param signal 信号
-     */
     public void setSignal(CommunitySignal signal) {
-        this.signal = EntityCopyUtils.copyCommunitySignal(signal);
+        this.signal = CommunityEntityCopyUtils.copyCommunitySignal(signal);
     }
 
-    /**
-     * 获取带有防御性副本的用户。
-     *
-     * @return 用户
-     */
     public User getUser() {
         return EntityCopyUtils.copyUser(user);
     }
 
-    /**
-     * 使用防御性副本设置用户。
-     *
-     * @param user 用户
-     */
     public void setUser(User user) {
         this.user = EntityCopyUtils.copyUser(user);
     }
