+++ koduck-backend/koduck-community/src/main/java/com/koduck/entity/community/UserSignalStats.java
@@ -18,6 +18,7 @@ import org.hibernate.annotations.UpdateTimestamp;
 
 import com.koduck.entity.auth.User;
 import com.koduck.util.EntityCopyUtils;
+import com.koduck.util.CommunityEntityCopyUtils;
 
 import lombok.AccessLevel;
 import lombok.Data;
