+++ koduck-backend/koduck-community/src/main/java/com/koduck/repository/community/SignalFavoriteRepository.java
